+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Change: Students can now delete entries until they're signed ([#358](https://gitlab.com/Skn0tt/EntE/issues/358))
+- Make deadline weak, such that only a warning is shown ([#354](https://gitlab.com/Skn0tt/EntE/issues/354))
 
 ## [v0.33.1] - 2020-04-01